Add files via upload
The dataset provides information on the performance of students. The data attributes include demographic, social and school related attributes which were collected using school reports and questionnaires. Interest is on factors that contributed to a student passing (as per their final grade). The aim of the project is to perform an appropriate two-way test of association for each of the relevant categorical explanatory variables and the response of interest. Clearly stating the hypotheses and conclusions. It also involves fitting a statistical model, finding the best fitting model and interpreting the results.